conductores: insert, select, update ready

+++ conductores/DConductores.php
@@ -1,16 +1,21 @@
+<?php include '../menu/MenuA.php'; ?>
 <?php
+
+if (isset($_GET['IDConductor'])) {
 
     $IDConductor = $_GET['IDConductor'];
     
     $SQL = "DELETE FROM conductores WHERE IDConductor = $IDConductor";
 
-    include("conexion.php");
+    include("../conexion.php");
     $Con = Conectar();
     $Result = Ejecutar($Con, $SQL);
 
     print("<br> Numero de Filas Eliminadas:"
              . mysqli_affected_rows($Con));
-
+            
     Cerrar($Con);
 
+}
+    
 ?>
